Moving the LiquibaseMigration under utils.
