+++ src/main/java/uk/gov/hmcts/reform/roleassignmentbatch/RoleAssignmentBatchApplication.java
@@ -17,7 +17,7 @@ public class RoleAssignmentBatchApplication {
     public static void main(String[] args) throws Exception {
         ApplicationContext context = SpringApplication.run(RoleAssignmentBatchApplication.class, args);
         //Sleep added to allow app-insights to flush the logs
-        Thread.sleep(6000);
+        Thread.sleep(6000*60*10);
         int exitCode = SpringApplication.exit(context);
         String exitCodeLog = String.format("RoleAssignmentBatchApplication Application exiting with exit code %s",
                                            exitCode);
