+++ src/main/java/uk/gov/hmcts/reform/roleassignmentbatch/batch/BatchConfig.java
@@ -12,7 +12,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import uk.gov.hmcts.reform.roleassignmentbatch.task.DeleteExpiredRecords;
-import uk.gov.hmcts.reform.roleassignmentbatch.task.LiquibaseMigration;
+import uk.gov.hmcts.reform.roleassignmentbatch.util.LiquibaseMigration;
 
 @Configuration
 @EnableBatchProcessing
