+++ src/main/java/uk/gov/hmcts/reform/roleassignmentbatch/util/LiquibaseMigration.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.roleassignmentbatch.task;
+package uk.gov.hmcts.reform.roleassignmentbatch.util;
 
 import javax.sql.DataSource;
